+++ backend/src/main/java/com/ssafy/keeping/domain/charge/dto/ssafyapi/request/SsafyAccountDepositRequestDto.java
@@ -21,7 +21,7 @@ public class SsafyAccountDepositRequestDto {
     public static SsafyAccountDepositRequestDto create(
             SsafyApiHeaderDto header,
             String accountNo,
-            BigDecimal transactionBalance,
+            Long transactionBalance,
             String transactionSummary) {
         
         return SsafyAccountDepositRequestDto.builder()
